+++ flutter/test/event_processor/flutter_enricher_event_processor_test.dart
@@ -74,7 +74,6 @@ void main() {
       final flutterContext = event?.contexts['flutter_context'];
       expect(flutterContext, isNotNull);
       expect(flutterContext, isA<Map<String, String>>());
-      expect(flutterContext['renderer'], isNotNull);
     });
 
     testWidgets('accessibility context', (WidgetTester tester) async {