+++ flutter/test/replay/replay_native_test.dart
@@ -42,10 +42,12 @@ void main() {
       setUp(() {
         hub = MockHub();
         native = NativeChannelFixture();
+
         options = defaultTestOptions()
           ..platform = mockPlatform
           ..methodChannel = native.channel
-          ..experimental.replay.quality = SentryReplayQuality.low;
+          ..replay.quality = SentryReplayQuality.low;
+
         sut = createBinding(options);
 
         if (mockPlatform.isIOS) {
@@ -72,8 +74,8 @@ void main() {
 
       group('replay recorder', () {
         setUp(() async {
-          options.experimental.replay.sessionSampleRate = 0.1;
-          options.experimental.replay.onErrorSampleRate = 0.1;
+          options.replay.sessionSampleRate = 0.1;
+          options.replay.onErrorSampleRate = 0.1;
           await sut.init(hub);
         });