+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Unreleased 9.0.0
+
+- Move replay and privacy from experimental to options ([#2755](https://github.com/getsentry/sentry-dart/pull/2755))
+- Remove renderer from `flutter_context` ([#2751](https://github.com/getsentry/sentry-dart/pull/2751))
+
 ## 9.0.0-alpha.1
 
 ### Breaking changes
@@ -27,6 +32,8 @@
   - The `PlatformChecker` was renamed to `RuntimeChecker`
   - Moved `PlatfomrChecker.platform` to `options.platform`
   
+
+
 ### Enhancements
 
 - Replay: improve Android native interop performance by using JNI ([#2670](https://github.com/getsentry/sentry-dart/pull/2670))